+++ src/WebLinterVsix/LinterService.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.IO;
+using System.Reflection;
 using System.Threading;
 using System.Windows.Threading;
 using WebLinter;
@@ -7,6 +9,8 @@ namespace WebLinterVsix
 {
     internal static class LinterService
     {
+        private static bool _defaultsCreated;
+
         static LinterService()
         {
             LinterFactory.Initializing += delegate { StatusText("Extracting latest version of the linters..."); };
@@ -20,6 +24,8 @@ namespace WebLinterVsix
                 try
                 {
                     ErrorList.CleanErrors(fileName);
+                    EnsureDefaults();
+
                     var result = LinterFactory.Lint(fileName);
 
                     ErrorListService.ProcessLintingResults(new[] { result });
@@ -37,6 +43,28 @@ namespace WebLinterVsix
             {
                 VSPackage.Dte.StatusBar.Text = message;
             }), DispatcherPriority.ApplicationIdle, null);
+        }
+
+        private static void EnsureDefaults()
+        {
+            if (!_defaultsCreated)
+            {
+                string assembly = Assembly.GetExecutingAssembly().Location;
+                string root = Path.GetDirectoryName(assembly);
+                string sourceFolder = Path.Combine(root, "Resources\\Defaults");
+                string destFolder = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
+
+                foreach (string sourceFile in Directory.EnumerateFiles(sourceFolder))
+                {
+                    string fileName = Path.GetFileName(sourceFile);
+                    string destFile = Path.Combine(destFolder, fileName);
+
+                    if (!File.Exists(destFile))
+                        File.Copy(sourceFile, destFile);
+                }
+
+                _defaultsCreated = true;
+            }
         }
     }
 }